Code cleanup/Eslint
Cleaned up the code, logic and component structure
