+++ ui/src/App.js
@@ -1,57 +1,20 @@
-import React from 'react';
-import './App.css';
-import { makeStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import Button from "@material-ui/core/Button";
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
+import React from 'react'
+import PropTypes from 'prop-types'
+import { useTheme } from '@material-ui/styles'
+import Navbar from './Components/Navbar/Navbar'
 
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  title: {
-    flexGrow: 1,
-  },
-}));
-
-export default function ButtonAppBar() {
-  const classes = useStyles();
-
+function App({ children }) {
   return (
-    <div className={classes.root}>
-      <AppBar position="static">
-        <Toolbar>
-          <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" className={classes.title}>
-            PathFinder
-          </Typography>
-          <Button color="inherit">FI</Button>
-          <Button color="inherit">SV</Button>
-          <Button color="inherit">ENG</Button>
-        </Toolbar>
-      </AppBar>
+    <div className='container'>
+      <Navbar />
+      <div className='children'>{children}</div>
     </div>
-  );
+  )
 }
 
-/* function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <h1>The App is here!</h1>
-      </header>
-    </div>
-  );
-} */
+App.propTypes = {
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
+  children: PropTypes.element.isRequired
+}
 
- 
-
+export default App
