--- ui/src/MyAwesomeReactComponent.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import RaisedButton from 'material-ui/RaisedButton';
- 
-const MyAwesomeReactComponent = () => (
-  <RaisedButton label="Default" />
-);
- 
-export default MyAwesomeReactComponent;
