+++ ui/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Demo from './demo';
+import App from './App.js';
 
-ReactDOM.render(<Demo />, document.querySelector('#root'));
+ReactDOM.render(<App />, document.querySelector('#root'));
